complete commentDAO and voteDAO

+++ server/VoteDAO.py
@@ -3,30 +3,31 @@ class VoteDAO:
         self.cursor = cursor
     
     def create(self, email, recipe, newVote):
-        request = "SELECT rating FROM Votes"
-        self.cursor.execute(request)
-        rating = self.cursor.fetchone()[0]
-        if rating:
+        request = "SELECT COUNT(*) FROM Votes WHERE email = %s AND recipe = %s"
+        self.cursor.execute(request, (email, recipe))
+        ratingCount = self.cursor.fetchone()[0]
+        if ratingCount > 0:
             return
-        else:
-            request = "INSERT INTO Votes(recipe, email, rating) values (%s, %s, %s)"
-            self.cursor.execute(request, (recipe, email, newVote))
+        request = "INSERT INTO Votes(recipe, email, rating) values (%s, %s, %s)"
+        self.cursor.execute(request, (recipe, email, newVote))
 
-    def edit(self, recipe, email, editVote):
-        request = "SELECT rating FROM Votes"
-        self.cursor.execute(request)
-        rating = self.cursor.fetchone()[0]
-        if rating:
+    def update(self, recipe, email, editVote):
+        request = "SELECT COUNT(*) FROM Votes WHERE email = %s AND recipe = %s"
+        self.cursor.execute(request, (email, recipe))
+        ratingCount = self.cursor.fetchone()[0]
+        if ratingCount > 0:
             request = "UPDATE Votes SET rating = %s WHERE recipe = %s AND email = %s"
             self.cursor.execute(request, (editVote, recipe, email))
-        else:
-            return
 
     def remove(self, recipe, email):
-        request = "DELETE FROM Votes WHERE recipe = %s AND email = %s"
-        self.cursor.execute(request, (recipe, email))
+        request = "SELECT COUNT(*) FROM Votes WHERE email = %s AND recipe = %s"
+        self.cursor.execute(request, (email, recipe))
+        ratingCount = self.cursor.fetchone()[0]
+        if ratingCount > 0:
+            request = "DELETE FROM Votes WHERE recipe = %s AND email = %s"
+            self.cursor.execute(request, (recipe, email))
 
-    def getFromUser(self, recipe, email):
-        request = "SELECT rating FROM Votes WHERE recipe = %s AND email = %s"
-        self.cursor.execute(request, (recipe, email))
+    def getFromUser(self, email):
+        request = "SELECT recipe, rating FROM Votes WHERE email = %s"
+        self.cursor.execute(request, (email))
         return self.cursor.fetchall()
